requirements.txt
streamlit
google-generativeai
python-dotenv
python-telegram-bot